# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LocalDramaStudio",
    "description": "LOCAL_ONLY local AI short-drama production workbench API",
    "version": "0.1.0-g1"
  },
  "paths": {
...
    "/api/v1/workflow-packages": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Register Package",
        "operationId": "registerWorkflowPackage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registerworkflowpackage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-packages:h3-candidate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Register H3 Candidate",
        "operationId": "registerH3CandidateWorkflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/H3CandidateWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registerh3Candidateworkflow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-packages:h3-i2v-candidate": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Register H3 I2V Candidate",
        "operationId": "registerH3I2VCandidateWorkflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/H3I2VCandidateWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registerh3I2Vcandidateworkflow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflow-versions/{version_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Version",
        "operationId": "getWorkflowVersion",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getworkflowversion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflow-versions/{version_id}:compile": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Compile Inputs",
        "operationId": "compileWorkflowInputs",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compileworkflowinputs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-versions/{version_id}:validate-local": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Validate Local",
        "operationId": "validateWorkflowLocal",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validateworkflowlocal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-versions/{version_id}:publish": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Publish",
        "operationId": "publishWorkflowVersion",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Publishworkflowversion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-versions/{version_id}:rollback": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Rollback",
        "operationId": "rollbackWorkflowVersion",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowPublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rollbackworkflowversion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow-versions/{version_id}:revoke": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Revoke",
        "operationId": "revokeWorkflowVersion",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Revokeworkflowversion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comfy/jobs:submit-next": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Submit Next",
        "operationId": "submitNextComfyJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComfyWorkerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Submitnextcomfyjob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comfy/attempts/{attempt_id}:poll": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Poll Attempt",
        "operationId": "pollComfyAttempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComfyWorkerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Pollcomfyattempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comfy/attempts/{attempt_id}:interrupt": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Interrupt Attempt",
        "operationId": "interruptComfyAttempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComfyWorkerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Interruptcomfyattempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comfy/attempts/{attempt_id}:recover": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Recover Attempt",
        "operationId": "recoverComfyAttempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComfyWorkerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recovercomfyattempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComfyWorkerRequest": {
        "properties": {
          "worker_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Worker Id"
          }
        },
        "type": "object",
        "required": [
          "worker_id"
        ],
        "title": "ComfyWorkerRequest"
      },
...
      "H3CandidateWorkflowRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed"
          },
          "duration_seconds": {
            "type": "number",
            "maximum": 15.0,
            "minimum": 4.0,
            "title": "Duration Seconds",
            "default": 4.0
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "16:9"
          },
          "filename_prefix": {
            "type": "string",
            "title": "Filename Prefix",
            "default": "local_drama/h3_candidate"
          },
          "sigma_points": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Sigma Points",
            "default": 50
          },
          "acceleration": {
            "type": "string",
            "title": "Acceleration",
            "default": "off"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "prompt",
          "seed"
        ],
        "title": "H3CandidateWorkflowRequest"
      },
      "H3I2VCandidateWorkflowRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed"
          },
          "duration_seconds": {
            "type": "number",
            "maximum": 15.0,
            "minimum": 4.0,
            "title": "Duration Seconds",
            "default": 4.0
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "auto"
          },
          "filename_prefix": {
            "type": "string",
            "title": "Filename Prefix",
            "default": "local_drama/h3_i2v_candidate"
          },
          "sigma_points": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Sigma Points",
            "default": 50
          },
          "acceleration": {
            "type": "string",
            "title": "Acceleration",
            "default": "off"
          },
          "first_frame": {
            "type": "string",
            "title": "First Frame",
            "default": "pending-approved-keyframe.png"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "prompt",
          "seed"
        ],
        "title": "H3I2VCandidateWorkflowRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowCompileRequest": {
        "properties": {
          "semantic_inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Semantic Inputs"
          }
        },
        "type": "object",
        "title": "WorkflowCompileRequest"
      },
      "WorkflowPackageRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "workflow": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workflow"
          },
          "contract": {
            "additionalProperties": true,
            "type": "object",
            "title": "Contract"
          },
          "node_bindings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Node Bindings"
          },
          "runtime_contract": {
            "additionalProperties": true,
            "type": "object",
            "title": "Runtime Contract"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "workflow"
        ],
        "title": "WorkflowPackageRequest"
      },
      "WorkflowPublishRequest": {
        "properties": {
          "validation_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Validation Id"
          }
        },
        "type": "object",
        "required": [
          "validation_id"
        ],
        "title": "WorkflowPublishRequest"
      }
    }
  }
}